add conditional rendering to pages, rename folders for better description
